Improve Creature class
Loads of code improvements in regards to const correctness, documentation, factoring common code into small helper functions and declaring variables closer to where they are needed. It appears there was a bug in CMOVE that assigned the move tick rate to the task's next_time after the creature attacked the player.