Modified meson to enable installation of files
Sound files and the dungeons-of-daggorath executable will now be installed to the correct folders in linux. Created a config.h.in file that holds the project version and the path to the sound files, meson will read this file and create a corresponding config.h file that can be used in the project.